chore: refactor extenders
